- make it compileable on 64-bit platforms and working with perl 5.7.1

+++ File.c
@@ -183,7 +183,7 @@ File_reset_notifications( Handle self)
 
    for ( i = 0; i < 3; i++) {
       if ( ret[i] == nil) continue;
-      list = var-> events + ( int) ret[i] - 1;
+      list = var-> events + ( IntPtr) ret[i] - 1;
       if ( list-> count > 0) mask |= cmd[ i];
    }
