+++ unix/apc_misc.c
@@ -782,7 +782,7 @@ apc_show_message( const char * message)
       apc_font_pick( nilHandle, &f, &f);
       cf = prima_find_known_font( &f, false, false);
       if ( !cf || !cf-> id) {
-         warn( "UAF_007: internal error (cf:%08x)", (uint)cf); /* the font was not cached, can't be */
+         warn( "UAF_007: internal error (cf:%08x)", (IntPtr)cf); /* the font was not cached, can't be */
          warn( message);
          return false;
       }
